apiserver/requirements.txt
mysql-connector==2.2.9
easierlog
flake8
mypy
flask
